+++ code/CapstoneTest/BackendTests/Model/TestWaypointManager/TestCreateWaypoint.cs
@@ -67,7 +67,7 @@ namespace CapstoneTest.BackendTests.Model.TestWaypointManager
             var resultResponse =
                 waypointManager.CreateWaypoint(1, "1601 Maple St", DateTime.Now, DateTime.Now.AddDays(2), null);
 
-            Assert.AreEqual(Ui.StatusCode.Success, resultResponse.StatusCode);
+            Assert.AreEqual((uint)Ui.StatusCode.Success, resultResponse.StatusCode);
         }
     }
 }
